split json_like.rs
