+++ src/json/json_like.rs
@@ -1,7 +1,5 @@
 use std::collections::HashMap;
 
-use async_graphql_value::ConstValue;
-
 pub trait JsonLike {
     type Output;
     fn as_array_ok(&self) -> Option<&Vec<Self::Output>>;
@@ -18,159 +16,6 @@ pub trait JsonLike {
     fn get_key(&self, path: &str) -> Option<&Self::Output>;
     fn from_output(value: &Self::Output) -> &Self;
     fn group_by<'a>(&'a self, path: &'a [String]) -> HashMap<String, Vec<&'a Self::Output>>;
-}
-
-impl JsonLike for serde_json::Value {
-    type Output = serde_json::Value;
-    fn as_array_ok(&self) -> Option<&Vec<Self::Output>> {
-        self.as_array()
-    }
-    fn as_str_ok(&self) -> Option<&str> {
-        self.as_str()
-    }
-    fn as_i64_ok(&self) -> Option<i64> {
-        self.as_i64()
-    }
-    fn as_u64_ok(&self) -> Option<u64> {
-        self.as_u64()
-    }
-    fn as_f64_ok(&self) -> Option<f64> {
-        self.as_f64()
-    }
-    fn as_bool_ok(&self) -> Option<bool> {
-        self.as_bool()
-    }
-    fn as_null_ok(&self) -> Option<()> {
-        self.as_null()
-    }
-
-    fn get_path<T: AsRef<str>>(&self, path: &[T]) -> Option<&Self::Output> {
-        let mut val = self;
-        for token in path {
-            val = match val {
-                serde_json::Value::Array(arr) => {
-                    let index = token.as_ref().parse::<usize>().ok()?;
-                    arr.get(index)?
-                }
-                serde_json::Value::Object(map) => map.get(token.as_ref())?,
-                _ => return None,
-            };
-        }
-        Some(val)
-    }
-
-    fn from_output(value: &Self::Output) -> &Self {
-        value
-    }
-
-    fn get_key(&self, path: &str) -> Option<&Self::Output> {
-        match self {
-            serde_json::Value::Object(map) => map.get(path),
-            _ => None,
-        }
-    }
-
-    fn as_string_ok(&self) -> Option<&String> {
-        match self {
-            serde_json::Value::String(s) => Some(s),
-            _ => None,
-        }
-    }
-
-    fn group_by<'a>(&'a self, path: &'a [String]) -> HashMap<String, Vec<&'a Self::Output>> {
-        let src = gather_path_matches(self, path, vec![]);
-        group_by_key(src)
-    }
-}
-
-impl JsonLike for async_graphql::Value {
-    type Output = async_graphql::Value;
-
-    fn as_array_ok(&self) -> Option<&Vec<Self::Output>> {
-        match self {
-            ConstValue::List(seq) => Some(seq),
-            _ => None,
-        }
-    }
-
-    fn as_str_ok(&self) -> Option<&str> {
-        match self {
-            ConstValue::String(s) => Some(s),
-            _ => None,
-        }
-    }
-
-    fn as_i64_ok(&self) -> Option<i64> {
-        match self {
-            ConstValue::Number(n) => n.as_i64(),
-            _ => None,
-        }
-    }
-
-    fn as_u64_ok(&self) -> Option<u64> {
-        match self {
-            ConstValue::Number(n) => n.as_u64(),
-            _ => None,
-        }
-    }
-
-    fn as_f64_ok(&self) -> Option<f64> {
-        match self {
-            ConstValue::Number(n) => n.as_f64(),
-            _ => None,
-        }
-    }
-
-    fn as_bool_ok(&self) -> Option<bool> {
-        match self {
-            ConstValue::Boolean(b) => Some(*b),
-            _ => None,
-        }
-    }
-
-    fn as_null_ok(&self) -> Option<()> {
-        match self {
-            ConstValue::Null => Some(()),
-            _ => None,
-        }
-    }
-
-    fn get_path<T: AsRef<str>>(&self, path: &[T]) -> Option<&Self::Output> {
-        let mut val = self;
-        for token in path {
-            val = match val {
-                ConstValue::List(seq) => {
-                    let index = token.as_ref().parse::<usize>().ok()?;
-                    seq.get(index)?
-                }
-                ConstValue::Object(map) => map.get(token.as_ref())?,
-                _ => return None,
-            };
-        }
-        Some(val)
-    }
-
-    fn from_output(value: &Self::Output) -> &Self {
-        value
-    }
-
-    fn get_key(&self, path: &str) -> Option<&Self::Output> {
-        match self {
-            ConstValue::Object(map) => map.get(&async_graphql::Name::new(path)),
-            _ => None,
-        }
-    }
-    fn as_string_ok(&self) -> Option<&String> {
-        match self {
-            ConstValue::String(s) => Some(s),
-            _ => None,
-        }
-    }
-
-    fn group_by<'a>(&'a self, path: &'a [String]) -> HashMap<String, Vec<&'a Self::Output>> {
-        let src = gather_path_matches(self, path, vec![]);
-        group_by_key(src)
-    }
 }
 
 // Highly micro-optimized and benchmarked version of get_path_all
@@ -197,7 +42,9 @@ pub fn gather_path_matches<'a, J: JsonLike>(
     vector
 }
 
-fn group_by_key<'a, J: JsonLike>(src: Vec<(&'a J, &'a J)>) -> HashMap<String, Vec<&'a J>> {
+pub(crate) fn group_by_key<'a, J: JsonLike>(
+    src: Vec<(&'a J, &'a J)>,
+) -> HashMap<String, Vec<&'a J>> {
     let mut map: HashMap<String, Vec<&'a J>> = HashMap::new();
     for (key, value) in src {
         // Need to handle number and string keys
@@ -215,119 +62,4 @@ fn group_by_key<'a, J: JsonLike>(src: Vec<(&'a J, &'a J)>) -> HashMap<String, Ve
         }
     }
     map
-}
-
-#[cfg(test)]
-mod tests {
-
-    use pretty_assertions::assert_eq;
-    use serde_json::json;
-
-    use super::*;
-
-    #[test]
-    fn test_gather_path_matches() {
-        let input = json!([
-            {"id": "1"},
-            {"id": "2"},
-            {"id": "3"}
-        ]);
-
-        let actual =
-            serde_json::to_value(gather_path_matches(&input, &["id".into()], vec![])).unwrap();
-
-        let expected = json!(
-            [
-              ["1", {"id": "1"}],
-              ["2", {"id": "2"}],
-              ["3", {"id": "3"}],
-            ]
-        );
-
-        assert_eq!(actual, expected)
-    }
-
-    #[test]
-    fn test_gather_path_matches_nested() {
-        let input = json!({
-            "data": [
-                {"user": {"id": "1"}},
-                {"user": {"id": "2"}},
-                {"user": {"id": "3"}},
-                {"user": [
-                    {"id": "4"},
-                    {"id": "5"}
-                    ]
-                },
-            ]
-        });
-
-        let actual = serde_json::to_value(gather_path_matches(
-            &input,
-            &["data".into(), "user".into(), "id".into()],
-            vec![],
-        ))
-        .unwrap();
-
-        let expected = json!(
-            [
-              ["1", {"id": "1"}],
-              ["2", {"id": "2"}],
-              ["3", {"id": "3"}],
-              ["4", {"id": "4"}],
-              ["5", {"id": "5"}],
-
-            ]
-        );
-
-        assert_eq!(actual, expected)
-    }
-
-    #[test]
-    fn test_group_by_key() {
-        let arr = vec![
-            (json!("1"), json!({"id": "1"})),
-            (json!("2"), json!({"id": "2"})),
-            (json!("2"), json!({"id": "2"})),
-            (json!("3"), json!({"id": "3"})),
-        ];
-        let input: Vec<(&serde_json::Value, &serde_json::Value)> =
-            arr.iter().map(|a| (&a.0, &a.1)).collect();
-
-        let actual = serde_json::to_value(group_by_key(input)).unwrap();
-
-        let expected = json!(
-            {
-                "1": [{"id": "1"}],
-                "2": [{"id": "2"}, {"id": "2"}],
-                "3": [{"id": "3"}],
-            }
-        );
-
-        assert_eq!(actual, expected)
-    }
-
-    #[test]
-    fn test_group_by_numeric_key() {
-        let arr = vec![
-            (json!(1), json!({"id": 1})),
-            (json!(2), json!({"id": 2})),
-            (json!(2), json!({"id": 2})),
-            (json!(3), json!({"id": 3})),
-        ];
-        let input: Vec<(&serde_json::Value, &serde_json::Value)> =
-            arr.iter().map(|a| (&a.0, &a.1)).collect();
-
-        let actual = serde_json::to_value(group_by_key(input)).unwrap();
-
-        let expected = json!(
-            {
-                "1": [{"id": 1}],
-                "2": [{"id": 2}, {"id": 2}],
-                "3": [{"id": 3}],
-            }
-        );
-
-        assert_eq!(actual, expected)
-    }
 }
