+++ src/json/mod.rs
@@ -1,4 +1,6 @@
+mod async_value;
 mod json_like;
 mod json_schema;
+mod serde_value;
 pub use json_like::*;
 pub use json_schema::*;
