+++ src/json/serde_value.rs
@@ -1,0 +1,181 @@
+use std::collections::HashMap;
+
+use super::{gather_path_matches, group_by_key, JsonLike};
+
+impl JsonLike for serde_json::Value {
+    type Output = serde_json::Value;
+    fn as_array_ok(&self) -> Option<&Vec<Self::Output>> {
+        self.as_array()
+    }
+    fn as_str_ok(&self) -> Option<&str> {
+        self.as_str()
+    }
+    fn as_i64_ok(&self) -> Option<i64> {
+        self.as_i64()
+    }
+    fn as_u64_ok(&self) -> Option<u64> {
+        self.as_u64()
+    }
+    fn as_f64_ok(&self) -> Option<f64> {
+        self.as_f64()
+    }
+    fn as_bool_ok(&self) -> Option<bool> {
+        self.as_bool()
+    }
+    fn as_null_ok(&self) -> Option<()> {
+        self.as_null()
+    }
+
+    fn get_path<T: AsRef<str>>(&self, path: &[T]) -> Option<&Self::Output> {
+        let mut val = self;
+        for token in path {
+            val = match val {
+                serde_json::Value::Array(arr) => {
+                    let index = token.as_ref().parse::<usize>().ok()?;
+                    arr.get(index)?
+                }
+                serde_json::Value::Object(map) => map.get(token.as_ref())?,
+                _ => return None,
+            };
+        }
+        Some(val)
+    }
+
+    fn from_output(value: &Self::Output) -> &Self {
+        value
+    }
+
+    fn get_key(&self, path: &str) -> Option<&Self::Output> {
+        match self {
+            serde_json::Value::Object(map) => map.get(path),
+            _ => None,
+        }
+    }
+
+    fn as_string_ok(&self) -> Option<&String> {
+        match self {
+            serde_json::Value::String(s) => Some(s),
+            _ => None,
+        }
+    }
+
+    fn group_by<'a>(&'a self, path: &'a [String]) -> HashMap<String, Vec<&'a Self::Output>> {
+        let src = gather_path_matches(self, path, vec![]);
+        group_by_key(src)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+
+    use pretty_assertions::assert_eq;
+    use serde_json::json;
+
+    use super::*;
+
+    #[test]
+    fn test_gather_path_matches() {
+        let input = json!([
+            {"id": "1"},
+            {"id": "2"},
+            {"id": "3"}
+        ]);
+
+        let actual =
+            serde_json::to_value(gather_path_matches(&input, &["id".into()], vec![])).unwrap();
+
+        let expected = json!(
+            [
+              ["1", {"id": "1"}],
+              ["2", {"id": "2"}],
+              ["3", {"id": "3"}],
+            ]
+        );
+
+        assert_eq!(actual, expected)
+    }
+
+    #[test]
+    fn test_gather_path_matches_nested() {
+        let input = json!({
+            "data": [
+                {"user": {"id": "1"}},
+                {"user": {"id": "2"}},
+                {"user": {"id": "3"}},
+                {"user": [
+                    {"id": "4"},
+                    {"id": "5"}
+                    ]
+                },
+            ]
+        });
+
+        let actual = serde_json::to_value(gather_path_matches(
+            &input,
+            &["data".into(), "user".into(), "id".into()],
+            vec![],
+        ))
+        .unwrap();
+
+        let expected = json!(
+            [
+              ["1", {"id": "1"}],
+              ["2", {"id": "2"}],
+              ["3", {"id": "3"}],
+              ["4", {"id": "4"}],
+              ["5", {"id": "5"}],
+
+            ]
+        );
+
+        assert_eq!(actual, expected)
+    }
+
+    #[test]
+    fn test_group_by_key() {
+        let arr = vec![
+            (json!("1"), json!({"id": "1"})),
+            (json!("2"), json!({"id": "2"})),
+            (json!("2"), json!({"id": "2"})),
+            (json!("3"), json!({"id": "3"})),
+        ];
+        let input: Vec<(&serde_json::Value, &serde_json::Value)> =
+            arr.iter().map(|a| (&a.0, &a.1)).collect();
+
+        let actual = serde_json::to_value(group_by_key(input)).unwrap();
+
+        let expected = json!(
+            {
+                "1": [{"id": "1"}],
+                "2": [{"id": "2"}, {"id": "2"}],
+                "3": [{"id": "3"}],
+            }
+        );
+
+        assert_eq!(actual, expected)
+    }
+
+    #[test]
+    fn test_group_by_numeric_key() {
+        let arr = vec![
+            (json!(1), json!({"id": 1})),
+            (json!(2), json!({"id": 2})),
+            (json!(2), json!({"id": 2})),
+            (json!(3), json!({"id": 3})),
+        ];
+        let input: Vec<(&serde_json::Value, &serde_json::Value)> =
+            arr.iter().map(|a| (&a.0, &a.1)).collect();
+
+        let actual = serde_json::to_value(group_by_key(input)).unwrap();
+
+        let expected = json!(
+            {
+                "1": [{"id": 1}],
+                "2": [{"id": 2}, {"id": 2}],
+                "3": [{"id": 3}],
+            }
+        );
+
+        assert_eq!(actual, expected)
+    }
+}
