+++ src/json/stringify.rs
@@ -1,0 +1,3 @@
+trait JsonStringify {
+
+}
